* Cleaned up simulator module

* Removed fine service interface

* refactored detetion services

* refactored model services

* removed util package

* cleaned up presentation layer

+++ simulator/src/main/java/be/kdg/simulator/services/impl/generators/FileGenerator.java
@@ -48,15 +48,16 @@ public class FileGenerator implements MessageGenerator {
         try (Scanner scanner = new Scanner(generatorConfig.getFilePath())) {
             final int NANO_SECONDS = 1000000;
             LocalDateTime localDateTimeForMessage = LocalDateTime.now();
-            int delay;
+            long delay;
 
             while (scanner.hasNext()) {
                 String line = scanner.nextLine();
                 String[] values = line.split(",");
 
-                delay = Integer.parseInt(values[2]);
+                delay = Long.parseLong(values[2]);
                 localDateTimeForMessage = localDateTimeForMessage.plusNanos(NANO_SECONDS * delay);
 
+                //create new object form line
                 if (Integer.parseInt(values[0]) <= generatorConfig.getMaxId()) {
                     cameraMessages.add(new CameraMessage(
                             Integer.parseInt(values[0]),
